implements exception handling

+++ src/main/java/com/example/pharmacyback/service/ProductService.java
@@ -1,9 +1,13 @@
 package com.example.pharmacyback.service;
 
+import com.example.pharmacyback.exceptions.ErrorMessages;
+import com.example.pharmacyback.exceptions.errors.ProductExistsException;
+import com.example.pharmacyback.exceptions.errors.ProductNotFoundException;
 import com.example.pharmacyback.model.Product;
 import com.example.pharmacyback.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 
@@ -17,51 +21,45 @@ public class ProductService {
     }
 
     public List<Product> getAllProducts() {
-        try {
-            return productRepository.findAll();
-        } catch (Exception e) {
-            return null;
-        }
+        return productRepository.findAll();
     }
 
     public Product getProductById(UUID uuid) {
-        try {
-            return productRepository.findProductById(uuid);
-        } catch (Exception e) {
-            return null;
-        }
+        return productRepository.findProductById(uuid)
+                .orElseThrow(() -> new ProductNotFoundException(ErrorMessages.PRODUCT_NOT_FOUND_EXCEPTION));
     }
 
     public List<Product> getTopFiveProductByPrice() {
-        try {
-            return productRepository.findAllOrderByPriceDesc();
-        } catch (Exception e) {
-            return null;
-        }
+        return productRepository.findAllOrderByPriceDesc();
     }
 
     public List<Product> getLeastFiveProductByPrice() {
-        try {
-            return productRepository.findAllOrderByPriceAsc();
-        } catch (Exception e) {
-            return null;
-        }
+        return productRepository.findAllOrderByPriceAsc();
     }
 
     public void deleteProduct(UUID id) {
-        try {
+        boolean prodExists = productRepository.existsById(id);
+        if(prodExists)
             productRepository.deleteById(id);
-        } catch (Exception e) {
-
-        }
+        else
+            throw new ProductNotFoundException(ErrorMessages.PRODUCT_NOT_FOUND_EXCEPTION);
     }
 
-    public Product saveProduct(Product product) {
-        try {
-            System.out.println(product);
+    public Product insertProduct(Product product) {
+        String name = product.getName();
+        Date expiryDate = product.getExpiryDate();
+        int countExisting = productRepository.countByNameAndExpiryDate(name, expiryDate);
+        if(countExisting == 0)
             return productRepository.save(product);
-        } catch (Exception e) {
-            return null;
-        }
+        else
+            throw new ProductExistsException(ErrorMessages.PRODUCT_EXISTS_EXCEPTION);
+    }
+
+    public Product updateProduct(Product product) {
+        boolean prodExists = productRepository.existsById(product.getId());
+        if(prodExists)
+            return productRepository.save(product);
+        else
+            throw new ProductNotFoundException(ErrorMessages.PRODUCT_NOT_FOUND_EXCEPTION);
     }
 }
